added dropdown minimizer and refactored the code by breaking out several components
