+++ src/app/checkout-page/page.tsx
@@ -1,4 +1,4 @@
-import Checkout from "./Checkout"
+import Checkout from "./components/Checkout"
 
 export default function CheckoutPage() {
     return (<Checkout />)
