+++ src/app/svgs/MinusIcon.tsx
@@ -1,0 +1,6 @@
+export default function MinusIcon() {
+    return (
+        <svg width="22" height="22" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M9.5 16.625C10.4357 16.625 11.3622 16.4407 12.2266 16.0826C13.0911 15.7246 13.8765 15.1998 14.5381 14.5381C15.1998 13.8765 15.7246 13.0911 16.0826 12.2266C16.4407 11.3622 16.625 10.4357 16.625 9.5C16.625 8.56433 16.4407 7.63783 16.0826 6.77338C15.7246 5.90894 15.1998 5.12348 14.5381 4.46186C13.8765 3.80025 13.0911 3.27542 12.2266 2.91736C11.3622 2.55929 10.4357 2.375 9.5 2.375C7.61033 2.375 5.79806 3.12567 4.46186 4.46186C3.12567 5.79806 2.375 7.61033 2.375 9.5C2.375 11.3897 3.12567 13.2019 4.46186 14.5381C5.79806 15.8743 7.61033 16.625 9.5 16.625ZM5.54167 10.2917H13.4583V8.70833H5.54167V10.2917Z" fill="black" />
+        </svg>)
+}
